refactor(measure): merged measure and cyclic-measure modules into a single measure module
