--- src/schema/cyclic-measure/__stubs__/index.ts
@@ -1,1 +1,0 @@
-export * from './cyclic-measure';
