+++ src/schema/measure/__mocks__/cyclic-measure.repository.ts
