+++ src/schema/measure/__stubs__/index.ts
@@ -1,1 +1,2 @@
 export * from './measure';
+export * from './cyclic-measure';
