+++ src/schema/measure/commands/create-cyclic-measure.spec.ts
