+++ src/schema/measure/commands/delete-cyclic-measure.command.ts
