+++ src/schema/measure/commands/delete-cyclic-measure.spec.ts
