+++ src/schema/measure/commands/update-cyclic-measure.command.ts
@@ -1,4 +1,4 @@
-import { UpdateCyclicMeasureInput } from 'src/schema/cyclic-measure/dto/update-cyclic-measure.input';
+import { UpdateCyclicMeasureInput } from 'src/schema/measure/dto/update-cyclic-measure.input';
 
 export class UpdateCyclicMeasureCommand {
 	public constructor(public readonly data: UpdateCyclicMeasureInput) {}
