+++ src/schema/measure/cyclic-measure.factory.spec.ts
