+++ src/schema/measure/cyclic-measure.repository.spec.ts
@@ -1,8 +1,8 @@
 import { MockedObjectDeep } from 'ts-jest';
 
 import { PrismaService } from '../../prisma.service';
-import { QuantityUnitOfMeasurement } from '../measure/types/measure';
 
+import { QuantityUnitOfMeasurement } from './types/measure';
 import { CyclicMeasureRepository } from './cyclic-measure.repository';
 import { deletedCyclicMeasure, domainCyclicMeasure, cyclicMeasureInput, updateCyclicMeasureInput } from './__stubs__';
 import { CyclicMeasureTypes } from './types/cyclic-measure';
