+++ src/schema/measure/cyclic-measure.repository.ts
