+++ src/schema/measure/dto/base-cyclic-measure.input.ts
@@ -2,7 +2,7 @@ import { Field, InputType } from '@nestjs/graphql';
 import { IsEnum, IsInt, IsOptional, IsUUID, MaxLength } from 'class-validator';
 
 import { CyclicMeasureTypes } from '../types/cyclic-measure';
-import { QuantityUnitOfMeasurement } from '../../measure/types/measure';
+import { QuantityUnitOfMeasurement } from '../types/measure';
 
 @InputType()
 export class BaseCyclicMeasureInput {
