+++ src/schema/measure/dto/create-cyclic-measure.input.ts
@@ -1,7 +1,7 @@
 import { Field, InputType } from '@nestjs/graphql';
 import { IsUUID } from 'class-validator';
 
-import { BaseCyclicMeasureInput } from '../dto/base-cyclic-measure.input';
+import { BaseCyclicMeasureInput } from './base-cyclic-measure.input';
 
 @InputType()
 export class CreateCyclicMeasureInput extends BaseCyclicMeasureInput {
