+++ src/schema/measure/queries/find-cyclic-measures.query.ts
