+++ src/schema/measure/types/cyclic-measure.repository.interface.ts
